+++ gui/src/renderer/app.tsx
@@ -3,7 +3,7 @@ import {
   push as pushHistory,
   replace as replaceHistory,
 } from 'connected-react-router';
-import { ipcRenderer, webFrame } from 'electron';
+import { ipcRenderer, shell, webFrame } from 'electron';
 import log from 'electron-log';
 import { createMemoryHistory } from 'history';
 import * as React from 'react';
@@ -286,6 +286,16 @@ export default class AppRenderer {
 
   public async removeAccountFromHistory(accountToken: AccountToken): Promise<void> {
     return IpcRendererEventChannel.accountHistory.removeItem(accountToken);
+  }
+
+  public async openLinkWithAuth(link: string): Promise<void> {
+    let token = '';
+    try {
+      token = await IpcRendererEventChannel.account.getWwwAuthToken();
+    } catch (e) {
+      log.error(`Failed to get the WWW auth token: ${e.message}`);
+    }
+    shell.openExternal(`${link}?token=${token}`);
   }
 
   public async setAllowLan(allowLan: boolean) {