+++ CHANGELOG.md
@@ -38,6 +38,10 @@ Line wrap the file at 100 chars.                                              Th
 - Add support for Android Lollipop.
 - Allow logging in without connectivity.
 
+### Changed
+- Account and WireGuard keys links in the App will now log the user in automatically.
+- Update FAQ URL to `https://mullvad.net/help/tag/mullvad-app/`
+
 ### Removed
 - Remove support for `MULLVAD_LOCALE` environment variable.